[CI] Collect all label rule sets in one module
Review follow-up: give maintainers a single place to read (and adjust)
everything that drives PR labeling. New label-rules.mjs holds all three
rule sets together — PATH_LABEL_RULES, PACKAGE_RULES, TYPE_RULES — each
with the notes that used to live next to its matcher.

The matcher modules (match-path-labels, rank-package-labels,
match-type-label) now import their rule set from label-rules.mjs and
default to it, so the runner just calls them with the PR facts. No
behavior change.
