+++ packages/meta/src/pr-labels/match-path-labels.mjs
@@ -2,6 +2,7 @@
 // resolves like any other import. The workflow installs it with `npm ci` here;
 // locally it comes from node_modules.
 import { minimatch } from 'minimatch';
+import { PATH_LABEL_RULES } from './label-rules.mjs';
 
 /**
  * Compute the path-based labels for a pull request: apply a label when any of
@@ -26,53 +27,3 @@ export function matchPathLabels(changedFiles, rules = PATH_LABEL_RULES) {
 	}
 	return labels;
 }
-
-// Path-based labels: the ones that map cleanly to paths AND have no count
-// limit — [Aspect], [Focus], [Feature], and [Type] Documentation. A wide
-// refactor can legitimately match many of them.
-//
-// [Package][...] labels are NOT here — rank-package-labels.mjs ranks packages
-// by lines changed and caps the count. [Type] labels other than Documentation
-// are NOT here either — match-type-label.mjs infers one from the PR title.
-//
-// This used to live in .github/labeler.yml (consumed by actions/labeler). Now
-// that we match globs ourselves, it lives here as plain data alongside the
-// other label rules — one source of truth, unit-testable, no YAML parsing.
-export const PATH_LABEL_RULES = {
-	// [Aspect] * — cross-cutting concerns inferable from paths.
-	'[Aspect] Browser': [
-		'packages/php-wasm/web/**',
-		'packages/php-wasm/web-service-worker/**',
-		'packages/playground/website/**',
-		'packages/playground/remote/**',
-	],
-	'[Aspect] Node.js': [
-		'packages/php-wasm/node/**',
-		'packages/php-wasm/cli/**',
-		'packages/playground/cli/**',
-	],
-	'[Aspect] Service Worker': [
-		'packages/php-wasm/web-service-worker/**',
-		'**/service-worker*.{ts,js}',
-	],
-	'[Aspect] Sqlite': ['**/sqlite*/**'],
-	'[Aspect] WordPress': [
-		'packages/playground/wordpress/**',
-		'packages/playground/wordpress-builds/**',
-	],
-	'[Aspect] Website': [
-		'packages/playground/website/**',
-		'packages/playground/website-extras/**',
-	],
-	// [Feature] / [Focus] — only the few that map cleanly to paths.
-	'[Feature] PHP.wasm': [
-		'packages/php-wasm/compile/**',
-		'packages/php-wasm/universal/**',
-	],
-	'[Focus] Developer Tools': [
-		'packages/playground/devtools-extension/**',
-		'packages/php-wasm/xdebug-bridge/**',
-	],
-	// [Type] Documentation — the one [Type] label with a clean path heuristic.
-	'[Type] Documentation': ['packages/docs/**', '**/*.md'],
-};
